🔌 (#7) - add use email mutation
